a few changes to improve unit test coverage

+++ src/typescript-generator/repository.js
@@ -1,3 +1,4 @@
+import { existsSync } from "node:fs";
 import fs from "node:fs/promises";
 import nodePath, { dirname } from "node:path";
 import { fileURLToPath } from "node:url";
@@ -49,12 +50,22 @@ export class Repository {
     }
   }
 
-  copyCoreFiles(destination) {
+  async copyCoreFiles(destination) {
+    const destinationPath = nodePath
+      .join(destination, "types.d.ts")
+      .replaceAll("\\", "/");
+
+    if (!existsSync(destinationPath)) {
+      return false;
+    }
+
+    await ensureDirectoryExists(destination);
+
     return fs.copyFile(
       nodePath
         .join(__dirname, "../../dist/server/types.d.ts")
         .replaceAll("\\", "/"),
-      nodePath.join(destination, "types.d.ts").replaceAll("\\", "/"),
+      destinationPath,
     );
   }
